requirements.txt
flask
pdf2docx
PyMuPDF
Pillow
gunicorn
